rag/requirements.txt
fastapi
uvicorn
chromadb
PyYAML
PyPDF2
python-pptx
python-docx
jaconv
transformers
pydantic
pydantic-settings
pytest
python-multipart
pycryptodome
loguru
pandas
openpyxl
pymupdf
rank_bm25
langchain==0.3.27
langchain-ollama==0.3.10
langchain-chroma==0.2.6
langchain-core==0.3.78
langchain-community==0.3.30
torch
langchain-huggingface==0.3.1
sentence-transformers 
huggingface_hub 
transformers 
accelerate
sudachipy
sudachidict_core
